DEV: admin functions, proxying, UI/UX
corrected most admin functions in backend to use SDK V3
Added a proxy+ integration to secure all endpoints with ease
and to make development flexible and less verbose in terraform
UI/UX improvements and got rid of most bootstrap css content

+++ frontend/src/app/administration/manage-admins/manage-admins.component.ts
@@ -3,6 +3,7 @@ import { AuthService } from '../../shared-modules/services/auth.service';
 import { AdminService } from '../services/admin.service';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { ToastrService } from 'ngx-toastr';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-manage-admins',
@@ -15,16 +16,18 @@ export class ManageAdminsComponent implements OnInit {
   public form: FormGroup;
 
   constructor(public auth: AuthService,
-              private adminService: AdminService,
-              private fb: FormBuilder,
-              private toastr: ToastrService) {
-    this.loadAdmins();
+    private adminService: AdminService,
+    private fb: FormBuilder,
+    private toastr: ToastrService,
+    private dg: MatDialog,
+  ) {
     this.form = this.fb.group({
       email: ['', [Validators.email, Validators.required, Validators.pattern(/[.]+[a-zA-Z0-9]+$/)]]
     });
   }
-
+  
   ngOnInit() {
+    this.loadAdmins();
   }
 
   loadAdmins() {
@@ -36,15 +39,22 @@ export class ManageAdminsComponent implements OnInit {
     });
   }
 
-  removeAdmin(uid: any) {
-    this.adminService.remove_admin(uid).then((res: any) => {
-      if (res.success) {
-        this.toastr.success('Admin removed successfully', 'Success');
-        this.loadAdmins();
-      } else {
-        this.toastr.error('Unable to remove admin, check email and try again', 'Error');
+  async removeAdmin(uid: any, name: string, email: string) {
+    const { AdminActionConfirmModalComponent } = await import('../modals/admin-action-confirm-modal/admin-action-confirm-modal.component');
+    const dialogRef = this.dg.open(AdminActionConfirmModalComponent, { data: {message: `Removing ${name} (${email})`}});
+
+    dialogRef.afterClosed().subscribe((yes) => {
+      if (yes) {
+        this.adminService.remove_admin(uid).then((res: any) => {
+          if (res.success) {
+            this.toastr.success('Admin removed successfully', 'Success');
+            this.loadAdmins();
+          } else {
+            this.toastr.error('Unable to remove admin, check email and try again', 'Error');
+          }
+        });
       }
-    });
+    })
   }
 
   createAdmin() {
